+++ src/pages/home.tsx
@@ -1,10 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import { Movie } from '../types/movie';
-import { Container, CircularProgress, Typography } from '@mui/material';
+import { Container, Typography } from '@mui/material';
 import { fetchPopularMovies } from '../endpoints/movie';
 import MovieListContainer from '../components/features/movie-list/movie-list-container';
 import SearchForm from '../components/features/search/search-form';
-import { filterMovies } from '../utils/functions';
 
 const HomePage: React.FC = () => {
   const [movies, setMovies] = useState<Movie[]>([]);
@@ -27,8 +26,6 @@ const HomePage: React.FC = () => {
 
     loadMovies();
   }, []);
-
-  if (loading) return <CircularProgress sx={{ display: 'block', margin: 'auto', mt: 5 }} />;
   if (error) return <Typography color="error" textAlign="center">{error}</Typography>;
 
   return (
@@ -37,7 +34,7 @@ const HomePage: React.FC = () => {
       <Typography variant="h4" textAlign="center" my={3}>
         {headerLabel}
       </Typography>
-      <MovieListContainer movies={movies}/>
+      <MovieListContainer movies={movies} loading={loading}/>
     </Container>
   );
 };